pyproject.toml
[tool.poetry]
name = "upnp-scrobbler"
version = "0.1.0"
description = "A LAST.fm scrobbler for UPnP players"
authors = ["GioF_71 <giovanni.fulco@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
xmltodict = "1.0.2"
requests = "2.32.5"
python-didl-lite = "1.4.1"
async-upnp-client = "0.45.0"
pylast = "5.4.0"
python-dotenv = "1.1.1"
platformdirs = "4.4.0"
subsonic-connector = "0.3.10"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
